型の間違え修正

+++ app/ConstParam.cpp
@@ -7,10 +7,10 @@ const float ConstParam::PID_TASK_90[3] = {0.5, 0.01, 0.05};
 const float ConstParam::PID_TASK_LOW[3] ={1.0, 0.01, 0.1};
 const float ConstParam::PID_TASK_80[3] = {1.40, 0.8, 0.1};
 
-const int ConstParam::SPEED_TASK_1 = 80;
-const int ConstParam::SPEED_TASK_2 = 80;
-const int ConstParam::SPEED_TASK_3 = 35;
-const int ConstParam::SPEED_TASK_90 = 95;
-const int ConstParam::SPEED_TASK_LOW = 20;
-const int ConstParam::SPEED_TASK_80 = 80;
+const short ConstParam::SPEED_TASK_1 = 80;
+const short ConstParam::SPEED_TASK_2 = 80;
+const short ConstParam::SPEED_TASK_3 = 35;
+const short ConstParam::SPEED_TASK_90 = 95;
+const short ConstParam::SPEED_TASK_LOW = 20;
+const short ConstParam::SPEED_TASK_80 = 80;
