+++ app/ConstParam.h
@@ -8,11 +8,11 @@ struct ConstParam
   static const float PID_TASK_80[3];
   static const float PID_TASK_OFF[3];
 
-  static const int SPEED_TASK_1;
-  static const int SPEED_TASK_2;
-  static const int SPEED_TASK_3;
-  static const int SPEED_TASK_90;
-  static const int SPEED_TASK_LOW;
-  static const int SPEED_TASK_80;
+  static const short SPEED_TASK_1;
+  static const short SPEED_TASK_2;
+  static const short SPEED_TASK_3;
+  static const short SPEED_TASK_90;
+  static const short SPEED_TASK_LOW;
+  static const short SPEED_TASK_80;
 
 };
